Extract winter.data.pagination
