+++ tests/controllers/simple_controller.py
@@ -6,8 +6,8 @@ from rest_framework.response import Response
 
 import winter
 from winter import ResponseEntity
-from winter.pagination import Page
-from winter.pagination import PagePosition
+from winter.data.pagination.page import Page
+from winter.data.pagination.page_position import PagePosition
 
 
 @dataclasses.dataclass
