+++ tests/pagination/test_page.py
@@ -1,5 +1,5 @@
-from winter.pagination import Page
-from winter.pagination import PagePosition
+from winter.data.pagination.page import Page
+from winter.data.pagination.page_position import PagePosition
 
 
 def test_iter_page():
