+++ tests/pagination/test_page_position_argument_inspector.py
@@ -2,9 +2,9 @@ import pytest
 from drf_yasg import openapi
 
 import winter
-from winter.pagination import PagePosition
-from winter.pagination import PagePositionArgumentResolver
-from winter.pagination import PagePositionArgumentsInspector
+from winter.data.pagination.page_position import PagePosition
+from winter.pagination.page_position_argument_inspector import PagePositionArgumentsInspector
+from winter.pagination.page_position_argument_resolver import PagePositionArgumentResolver
 from winter.routing import get_route
