+++ tests/pagination/test_page_position_argument_resolver.py
@@ -7,9 +7,9 @@ from rest_framework.request import Request as DRFRequest
 import winter
 from winter import converters
 from winter.core import ComponentMethod
-from winter.pagination import PagePosition
-from winter.pagination import PagePositionArgumentResolver
-from winter.pagination import Sort
+from winter.data.pagination.page_position import PagePosition
+from winter.data.pagination.sort import Sort
+from winter.pagination.page_position_argument_resolver import PagePositionArgumentResolver
 
 
 @pytest.mark.parametrize(
