+++ winter/pagination/page/page_processor.py
@@ -2,9 +2,9 @@ import typing
 
 from rest_framework.request import Request as DRFRequest
 
-from .page import Page
 from ..utils import get_next_page_url
 from ..utils import get_previous_page_url
+from ...data.pagination.page import Page
 from ...output_processor import IOutputProcessor
