+++ winter/pagination/page/page_processor_resolver.py
@@ -1,7 +1,7 @@
-from typing import Any
+import typing
 
-from .page import Page
 from .page_processor import PageProcessor
+from ...data.pagination.page import Page
 from ...output_processor import IOutputProcessorResolver
 
 
@@ -10,8 +10,8 @@ class PageOutputProcessorResolver(IOutputProcessorResolver):
     def __init__(self):
         self._page_processor = PageProcessor()
 
-    def is_supported(self, body: Any) -> bool:
+    def is_supported(self, body: typing.Any) -> bool:
         return isinstance(body, Page)
 
-    def get_processor(self, body: Any) -> PageProcessor:
+    def get_processor(self, body: typing.Any) -> PageProcessor:
         return self._page_processor
