+++ winter/pagination/page_position_argument_inspector.py
@@ -1,14 +1,13 @@
-from typing import List
-from typing import TYPE_CHECKING
+import typing
 
 from drf_yasg import openapi
 
-from .page_position import PagePosition
 from .page_position_argument_resolver import PagePositionArgumentResolver
-from .sort import OrderByAnnotation
+from .sort.order_by import OrderByAnnotation
+from ..data.pagination.page_position import PagePosition
 from ..schema import MethodArgumentsInspector
 
-if TYPE_CHECKING:
+if typing.TYPE_CHECKING:
     from ..routing import Route
 
 
@@ -30,7 +29,7 @@ class PagePositionArgumentsInspector(MethodArgumentsInspector):
             type=openapi.TYPE_INTEGER,
         )
 
-    def inspect_parameters(self, route: 'Route') -> List[openapi.Parameter]:
+    def inspect_parameters(self, route: 'Route') -> typing.List[openapi.Parameter]:
         parameters = []
         has_page_position_argument = any(argument.type_ == PagePosition for argument in route.method.arguments)
         if not has_page_position_argument:
