+++ winter/pagination/sort/order_by.py
@@ -1,4 +1,5 @@
-import typing
+from typing import Iterable
+from typing import Tuple
 
 from .check_sort import check_sort
 from .order_by_annotation import OrderByAnnotation
@@ -6,7 +7,7 @@ from .parse_sort import parse_sort
 from ...core.annotation_decorator import annotate_method
 
 
-def order_by(allowed_fields: typing.Iterable[str], default_sort: typing.Tuple[str] = None):
+def order_by(allowed_fields: Iterable[str], default_sort: Tuple[str] = None):
     allowed_fields = frozenset(allowed_fields)
     if default_sort is not None:
         default_sort = parse_sort(','.join(default_sort))
