+++ winter/pagination/sort/order_by_annotation.py
@@ -2,11 +2,10 @@ import typing
 
 import dataclasses
 
-if typing.TYPE_CHECKING:
-    from .sort import Sort  # noqa: F401
+from ...data.pagination.sort import Sort
 
 
 @dataclasses.dataclass
 class OrderByAnnotation:
     allowed_fields: typing.FrozenSet[str]
-    default_sort: typing.Optional['Sort'] = None
+    default_sort: typing.Optional[Sort] = None
