+++ winter/pagination/sort/parse_sort.py
@@ -1,7 +1,7 @@
 import typing
 
 from .parse_order import parse_order
-from .sort import Sort
+from ...data.pagination.sort import Sort
 
 
 def parse_sort(str_sort: typing.Optional[str]) -> typing.Optional[Sort]:
