+++ winter/pagination/utils.py
@@ -4,7 +4,7 @@ from rest_framework.request import Request as DRFRequest
 from rest_framework.utils.urls import remove_query_param
 from rest_framework.utils.urls import replace_query_param
 
-from .page import Page
+from ..data.pagination.page import Page
 
 
 def get_previous_page_url(page: Page, request: DRFRequest) -> Optional[str]:
